de-22.0.54157.54635

+++ BaseApp/Source/Base Application/TransferExtendedText.Codeunit.al
@@ -29,7 +29,7 @@ codeunit 378 "Transfer Extended Text"
         ExtTextHeader: Record "Extended Text Header";
     begin
         MakeUpdateRequired := false;
-        if IsDeleteAttachedLines(SalesLine."Line No.", SalesLine."No.", SalesLine."Attached to Line No.") then
+        if IsDeleteAttachedLines(SalesLine."Line No.", SalesLine."No.", SalesLine."Attached to Line No.") and not SalesLine.IsExtendedText() then
             MakeUpdateRequired := DeleteSalesLines(SalesLine);
 
         AutoText := false;
@@ -158,7 +158,7 @@ codeunit 378 "Transfer Extended Text"
         ExtTextHeader: Record "Extended Text Header";
     begin
         MakeUpdateRequired := false;
-        if IsDeleteAttachedLines(PurchLine."Line No.", PurchLine."No.", PurchLine."Attached to Line No.") then
+        if IsDeleteAttachedLines(PurchLine."Line No.", PurchLine."No.", PurchLine."Attached to Line No.") and not PurchLine.IsExtendedText() then
             MakeUpdateRequired := DeletePurchLines(PurchLine);
 
         AutoText := false;
@@ -252,7 +252,9 @@ codeunit 378 "Transfer Extended Text"
     var
         ToSalesLine: Record "Sales Line";
         IsHandled: Boolean;
+#if not CLEAN22
         CompatibilityMode: Boolean;
+#endif
     begin
         OnBeforeInsertSalesExtText(SalesLine, TempExtTextLine, IsHandled, MakeUpdateRequired, LastInsertedSalesLine);
         if IsHandled then
@@ -265,6 +267,7 @@ codeunit 378 "Transfer Extended Text"
         ToSalesLine.SetRange("Document No.", SalesLine."Document No.");
         ToSalesLine := SalesLine;
         OnInsertSalesExtTextRetLastOnBeforeToSalesLineFind(ToSalesLine);
+#if not CLEAN22
         CompatibilityMode := false; // Disable previous LineSpacing assignment
         OnInsertSalesExtTextRetLastOnBeforeSetCompatibilityMode(CompatibilityMode);
         If CompatibilityMode then
@@ -276,6 +279,7 @@ codeunit 378 "Transfer Extended Text"
                     Error(Text000);
             end else
                 LineSpacing := 10000;
+#endif
 
         NextLineNo := SalesLine."Line No." + LineSpacing;
 
@@ -881,9 +885,12 @@ codeunit 378 "Transfer Extended Text"
     begin
     end;
 
+#if not CLEAN22
+    [Obsolete('Temporary event to support enabling of compatibility mode', '22.0')]
     [IntegrationEvent(false, false)]
     local procedure OnInsertSalesExtTextRetLastOnBeforeSetCompatibilityMode(var CompatibilityMode: Boolean)
     begin
     end;
+#endif
 }
